Zone Neighbor Dynamics
Zones now cannot have duplicate neighbors. Neighbors that do not exist as a zone will be created as one and added to zone list and selected zone's neighbors.  Neighbors that exist in the zone list will be added to the selected zone's neighbors without creating a new zone. The removal of the selected zone as a neighbor to a removed neighbor is now optional.