Save Files Implemented
The generator can now save and open a text file of zones created with the file type extension of ".ctnt". The file format needs to be updated to a string that would likely never be used in an entry's name. Currently set to ', ' and test input need to avoid having this separator for now.